fix(gfql/index): share array protocol typing
